Add an option in totalmass diagnostic to normalize the air mass.
I.e., scale the tracer mass as if the total air mass was constant throughout
the experiment.

This gets rid of the daily "wiggles" that have nothing to do with changing
tracer mass, but are instead due to the changing total air mass from the
surface pressure analyses.

Note: this implementation is somewhat fragile - it assumes that surface
pressure and 3D tracer outputs are at the same time frequencies.