Modify the Cache to accept a list of multiple save/load hooks.
Also, need to modify the existing hooks to read a dataset and return a dataset.

Before, the save hook would convert a single var to a dataset, and the load
hook would convert a dataset to a single var.  This can't be done if we allow
multiple hooks to be chained together, though.