Catch any ImportErrors when opening an existing manifest file.
Re-scan the source files if such an error occurs (indicates the structure of
the code has changed since the diagnostics were last run, so we can't use the
old references anymore).

Needed to do this for my tests because I just moved the gem interface to the
'products' subfolder, so the 'gem' reference in the manifest was invalid.