Split the GEM code into a core GEM interface, and an EC-CAS interface (for dealing with EC-CAS related tracers).
Move the methods into a class, to simplify inheriting from a general interface
to a more specific one (e.g. allow EC-CAS to inherit some default logic from
the GEM interface).

Remove the flag for dry air in the EC-CAS interface.  This information will
need to be detectable from the files themselves, not something the user has to
manually set.