Have a version of VertRegrid that seems to actually work.
Had to fix a bug with the caching mechanism - the same cache was being shared
by all instances of VertRegrid data, so the output fields all had the same
data!

Note: Need to do something about the diagnostic level, which has no dp extent.
It has NaN values, which are messing with xrec.