Implement a first version of horizontal regridding.
Also, make sure the wrap the outputs of horizontal and vertical regridding
into DataInterface objects.

Also, in common.rotate_grid(), check if the grid is already rotated, and don't
do anything.  This was causing a crash when the horizontal regridding was
turned on, because the first longitude of GEM's (already rotated) grid was
slightly negative, which punted it to the other end of the domain when doing
the rotation again.