For diagnostics on CO2 components (fossil, etc.), make sure the species name is identified.
Otherwise, the unit convertor won't know how to convert between units of mass
and moles.

Also, don't force the convertor to use the field name as the context of
conversion, since this won't work with components (e.g. the convertor would be
looking for mass of "CO2_fossil" molecules instead of "CO2" molecules).