When converting from a dry mass mixing ratio to a moist mass mixing ratio (for computing mass), explicitly carry over the specie name.
Otherwise, there is not enough context to convert to units of 'kg kg(air)-1',
since the convertor doesn't know what the molecule in the numerator is supposed
to be.

This used to work, because a previous version of the PyGeode package would
propogate the specie name when we compute c*(1-q), since q didn't have a specie
name of its own.  This was considered a bug (can't assume we can keep the same
attributes for the result), and was recently corrected.  This bugfix caused our
total mass diagnostic to fail.