Change units for CarbonTracker and GEOS-Chem to be molefractions of "semi-dry" air.
The specific context of when this means will depend on the diagnostic being
performed (e.g. "semi-dry" is "dry" when looking at molefractions, but "moist"
when looking at mass).

Addresses inconsistency in how some models handle their units (they may assume
a dry-air molefraction when comparing to obs, but then use the full surface
pressure field when doing a mass budget (such as handling fluxes?).

Note: Need to confirm this is what they actually do, I'm just basing this on
the results we get when we run diagnostics on their products.