Workaround a problem with caching of time-averaged data (e.g. monthly mean).
PyGeode doesn't write enough metadata to netcdf regarding the time axis, so
when it reads it back in, it adds a 'day' property and makes the data
impossible to work with.

As a quick fix, force the time axis to be what we expect it to be, and ignore
the time axis reported from the cache file.