Refactor TimeSeriesRBP to inherit from Timeseries.
Note: tried to keep it consistent with the previous behaviour, e.g. model is
not sampled at obs times.

Count is for the total number of stations, regardless of whether or not there
was any obs data available.

Aside: this diagnostic hadn't run for a while, because the obs dataset it was
looking for (GAW-2014) had been renamed to GAW-2014-hourly.