Add some support for products with multiple datasets.
Account for case where there may be different species measured at different
stations, so may have multiple domains because of that reason.  Use the domain
that maximizes the amount of data for the required specie.

Still no support for the case of splitting each "station" into a different
domain (e.g. for aircraft).  Which is okay, because this module is only for
surface station data anyway.