Remove all code dealing with "common" axes.
This was used for forcing all data (e.g. from obs datasets) to all use the
same time axis.  The only reason to do this was to make it easier to cache the
obs data all at once, and pass it around as a single coherent dataset.

However, forcing a common axis uses a lot of CPU cycles (especially for long-
running obs datasets).