Fix the logic for the --no-require-obs flag.
Instead of disabling the require_fieldname flag on all inputs, selectivly
change the behaviour for the case where we have no applicable obs data but
want to sample the model anyway.

This fixes a bug when using --no-require-obs on fields that are available
somehwere in the obs, but not at all stations.  Was getting empty datasets
for those locations, which was wreaking havoc with the timeseries diags.