For zonal-std-of-diff, do the diff before the zonal std, but *after* the vertical interpolation.
That way, each input field is interpolated using its own geopotential height.

Previously, the diff was done on the 3D field, the geopotential height from
the first input was copied to the diff, then it was interpolated.  This could
cause some small errors if the geopotential height was different for the two
experiments.