Add a --fields argument to eccas-regrid.
I needed this to regrid the new GEOS-Chem CO fluxes.  That data product also
included initial condtions, which the regridder was trying to vertically
interpolate (even though that field would be discarded when writing the
output).  There was no surface pressure field in the data, so the regridder
crashed.  This new argument allowed me to exclude those troublesome 3D fields.